+++ corehq/apps/hqwebapp/utils.py
@@ -8,9 +8,11 @@ from Crypto.Signature import PKCS1_PSS
 from django.templatetags.i18n import language_name
 
 from dimagi.utils.decorators.memoized import memoized
+from dimagi.utils.logging import notify_exception
 from corehq.apps.hqwebapp.forms import BulkUploadForm
 from corehq.apps.hqwebapp.tasks import send_html_email_async
 from corehq.apps.users.models import WebUser
+from corehq.util.view_utils import get_request
 from custom.nic_compliance.utils import get_raw_password
 
 logger = logging.getLogger(__name__)
@@ -88,3 +90,36 @@ def decode_password(obfuscated_password, username=None):
         return get_raw_password(obfuscated_password, username)
     else:
         return obfuscated_password
+
+
+def format_angular_error(error_msg, log_error):
+    """Gets the standard angular async error response.
+    :param error_msg: A string that is the error message you'd like to return
+    :param additional_data: a dictionary of additional data you'd like to pass
+    :return: {
+        'error': <error_msg>,
+        <...additional_data...>,
+    }
+    """
+    resp = {'error': error_msg}
+
+    if log_error:
+        notify_exception(get_request(), error_msg)
+
+    return resp
+
+
+def format_angular_success(additional_data=None):
+    """Gets the standard angular async SUCCESS response.
+    :param additional_data: a dictionary of additional data you'd like to pass
+    :return: {
+        'success': True,
+        <...additional_data...>,
+    }
+    """
+    resp = {
+        'success': True,
+    }
+    if isinstance(additional_data, dict):
+        resp.update(additional_data)
+    return resp
